FIX: all target types should be float32 to allow MTL training

+++ finetune/target_models/multiple_choice.py
@@ -41,7 +41,7 @@ class MultipleChoicePipeline(BasePipeline):
     def feed_shape_type_def(self):
         TS = tf.TensorShape
         return (
-            ({"tokens": tf.int32, "mask": tf.float32}, tf.int32),
+            ({"tokens": tf.int32, "mask": tf.float32}, tf.float32),
             (
                 {
                     "tokens": TS([self.target_dim, self.config.max_length, 2]),
